requirements.txt
os
google.cloud
io
streamlit